backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
numpy==1.24.3
opencv-python==4.8.1.78
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
requests==2.31.0
googlemaps==4.10.0
pillow==10.1.0
rasterio==1.3.9
sentinelsat==1.2.1
pymongo==4.6.1
